Moved DBFilterBase out of FindOperation so it can be use in other operations like count
